Fix incorrect package MD5sums being generated on amd64; incorrect data types used for 64 bit machine. Amended to be ISO standard datatypes. autobuilder/toolsrc/libpkg/libpkg/md5.h
git-svn-id: svn://svn.riscos.info/gccsdk/trunk@5971 5d7976f4-0b0e-0410-a586-d2bac8e3be52

+++ autobuilder/toolsrc/libpkg/libpkg/md5.h
@@ -9,6 +9,7 @@
 #include <cstddef>
 #include <string>
 #include <iosfwd>
+#include <stdint.h>
 
 namespace pkg {
 
@@ -23,10 +24,10 @@ class md5
 {
 public:
 	/** A type to represent an unsigned 32-bit integer. */
-	typedef unsigned long uint32;
+	typedef uint32_t uint32;
 
 	/** A type to represent an unsigned 64-bit integer. */
-	typedef unsigned long long uint64;
+	typedef uint64_t uint64;
 private:
 	/** The size of the message digest state, in 32-bit words. */
 	static const int state_size=4;
